Add: All Apis of Order

+++ UnicoVehicle/UnicoVehicle.DTO/Order.cs
@@ -8,7 +8,7 @@ namespace UnicoVehicle.DTO
         public Showroom Showroom { get; set; }
         public Status OrderStatus { get; set; }
         public Customer Customer { get; set; }
-        public float FinalPrice { get; set; }
+        public double FinalPrice { get; set; }
         public bool hasLoan { get; set; }
         public DateTime DeliveryDate { get; set; }
     }
